requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==45.0.5
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.7.0
duplicity==0.8.12.0
ecdsa==0.19.1
email_validator==2.2.0
entrypoints==0.3
exceptiongroup==1.3.0
fastapi==0.115.13
fasteners==0.14.1
future==0.18.2
greenlet==3.2.3
h11==0.16.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.3.10
MarkupSafe==3.0.2
monotonic==1.5
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.9.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
systemd-python==234
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.34.3
wadllib==1.3.3
xkit==0.0.0
